// frontend/src/api/model/getAvailableProductsToSellDto.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Luxis API
 * System API for Luxis
 * OpenAPI spec version: 1.0
 */
import type { GetAvailableCategoryDto } from './getAvailableCategoryDto'

export interface GetAvailableProductsToSellDto {
  /** The available categories with their product models and products */
  data: GetAvailableCategoryDto[]
}